DOOM-0140: fix remaining -Wall warnings (casts, buffers, types) -> 0
Warning sweep part 2 (no behaviour change); make is now warning-free (was 98):
- 64-bit pointer<->int casts via intptr_t/uintptr_t: p_saveg PADSAVEP +
  load-side index casts, z_zone owner-pointer check, d_main statcopy,
  d_net offsetof-style hacks. (#include <stdint.h> where needed.)
- i_net: recvfrom address-length is socklen_t, not int (pointer-sign).
- sprintf targets enlarged for the format's worst case: f_finale name[16],
  p_setup lumpname[16], wi_stuff name[40].
- g_game: strcmp save_p cast to (char*) (pointer-sign).
- r_things: the vissprite sort sentinel 'unsorted' is static, so its address
  isn't a dangling local (false-positive -Wdangling-pointer).

Two latent VANILLA bugs were found and PRESERVED (silenced, not fixed), each
parenthesised/cast to keep shipped behaviour and tracked for a decision:
DOOM-0138 (EV_DoDonut !/& precedence) and DOOM-0139 (sky gamemode vs
gamemission enum compare).

+++ linuxdoom-1.10/d_main.c
@@ -33,6 +33,7 @@ static const char rcsid[] __attribute__((used)) = "$Id: d_main.c,v 1.8 1997/02/0
 
 #ifdef NORMALUNIX
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
@@ -1239,7 +1240,7 @@ void D_DoomMain (void)
 	// for statistics driver
 	extern  void*	statcopy;                            
 
-	statcopy = (void*)atoi(myargv[p+1]);
+	statcopy = (void*)(intptr_t)atoi(myargv[p+1]);
 	printf ("External statistics registered.\n");
     }
